refactor: move monitoring related structs to monitoring mod
