+++ state-chain/custom-rpc/src/lib.rs
@@ -17,7 +17,6 @@
 use crate::{
 	backend::{CustomRpcBackend, NotificationBehaviour},
 	boost_pool_rpc::BoostPoolFeesRpc,
-	monitoring::RpcEpochState,
 };
 use boost_pool_rpc::BoostPoolDetailsRpc;
 use cf_amm::{
@@ -27,9 +26,7 @@ use cf_amm::{
 };
 use cf_chains::{
 	address::{AddressString, ForeignChainAddressHumanreadable, ToHumanreadableAddress},
-	dot::PolkadotAccountId,
 	eth::Address as EthereumAddress,
-	sol::SolAddress,
 	CcmChannelMetadata, Chain, VaultSwapExtraParametersRpc, MAX_CCM_MSG_LENGTH,
 };
 use cf_primitives::{
@@ -70,11 +67,6 @@ use sp_state_machine::InspectState;
 use state_chain_runtime::{
 	chainflip::{BlockUpdate, Offence},
 	constants::common::TX_FEE_MULTIPLIER,
-	monitoring_apis::{
-		ActivateKeysBroadcastIds, AuthoritiesInfo, BtcUtxos, ExternalChainsBlockHeight,
-		FeeImbalance, FlipSupply, LastRuntimeUpgradeInfo, MonitoringDataV2, OpenDepositChannels,
-		PendingBroadcasts, PendingTssCeremonies, RedemptionsInfo, SolanaNonces,
-	},
 	runtime_apis::{
 		AuctionState, BoostPoolDepth, BoostPoolDetails, BrokerInfo, CcmData, ChainAccounts,
 		CustomRuntimeApi, DispatchErrorWithMessage, ElectoralRuntimeApi, FailingWitnessValidators,
@@ -93,79 +85,6 @@ use std::{
 pub mod backend;
 pub mod monitoring;
 pub mod order_fills;
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct RpcRedemptionsInfo {
-	pub total_balance: NumberOrHex,
-	pub count: u32,
-}
-impl From<RedemptionsInfo> for RpcRedemptionsInfo {
-	fn from(redemption_info: RedemptionsInfo) -> Self {
-		Self { total_balance: redemption_info.total_balance.into(), count: redemption_info.count }
-	}
-}
-
-pub type RpcFeeImbalance = FeeImbalance<NumberOrHex>;
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct RpcFlipSupply {
-	pub total_supply: NumberOrHex,
-	pub offchain_supply: NumberOrHex,
-}
-impl From<FlipSupply> for RpcFlipSupply {
-	fn from(flip_supply: FlipSupply) -> Self {
-		Self {
-			total_supply: flip_supply.total_supply.into(),
-			offchain_supply: flip_supply.offchain_supply.into(),
-		}
-	}
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct RpcMonitoringData {
-	pub external_chains_height: ExternalChainsBlockHeight,
-	pub btc_utxos: BtcUtxos,
-	pub epoch: RpcEpochState,
-	pub pending_redemptions: RpcRedemptionsInfo,
-	pub pending_broadcasts: PendingBroadcasts,
-	pub pending_tss: PendingTssCeremonies,
-	pub open_deposit_channels: OpenDepositChannels,
-	pub fee_imbalance: RpcFeeImbalance,
-	pub authorities: AuthoritiesInfo,
-	pub build_version: LastRuntimeUpgradeInfo,
-	pub suspended_validators: Vec<(Offence, u32)>,
-	pub pending_swaps: u32,
-	pub dot_aggkey: PolkadotAccountId,
-	pub flip_supply: RpcFlipSupply,
-	pub sol_aggkey: SolAddress,
-	pub sol_onchain_key: SolAddress,
-	pub sol_nonces: SolanaNonces,
-	pub activating_key_broadcast_ids: ActivateKeysBroadcastIds,
-}
-impl From<MonitoringDataV2> for RpcMonitoringData {
-	fn from(monitoring_data: MonitoringDataV2) -> Self {
-		Self {
-			epoch: monitoring_data.epoch.into(),
-			pending_redemptions: monitoring_data.pending_redemptions.into(),
-			fee_imbalance: monitoring_data.fee_imbalance.map(|i| (*i).into()),
-			external_chains_height: monitoring_data.external_chains_height,
-			btc_utxos: monitoring_data.btc_utxos,
-			pending_broadcasts: monitoring_data.pending_broadcasts,
-			pending_tss: monitoring_data.pending_tss,
-			open_deposit_channels: monitoring_data.open_deposit_channels,
-			authorities: monitoring_data.authorities,
-			build_version: monitoring_data.build_version,
-			suspended_validators: monitoring_data.suspended_validators,
-			pending_swaps: monitoring_data.pending_swaps,
-			dot_aggkey: monitoring_data.dot_aggkey,
-			flip_supply: monitoring_data.flip_supply.into(),
-			sol_aggkey: monitoring_data.sol_aggkey,
-			sol_onchain_key: monitoring_data.sol_onchain_key,
-			sol_nonces: monitoring_data.sol_nonces,
-			activating_key_broadcast_ids: monitoring_data.activating_key_broadcast_ids,
-		}
-	}
-}
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ScheduledSwap {
